fix initialization
