+++ src/EnergyPlus/HeatBalanceManager.hh
@@ -186,11 +186,6 @@ namespace HeatBalanceManager {
 
     void InitConductionTransferFunctions(EnergyPlusData &state);
 
-    void EvalOutsideMovableInsulation(EnergyPlusData &state);
-
-    void EvalInsideMovableInsulation(EnergyPlusData &state);
-
-
 } // namespace HeatBalanceManager
 
 struct HeatBalanceMgrData : BaseGlobalStruct
